fix: restore required features hidden behind cffi
In my haste to merge the `cffi` feature, I was too aggressive in what
needed to move behind the flag. This broke my work in `flux-lsp`, which
doesn't use the `cffi` feature. This patch restores what shouldn't be
behind a feature flag.
