+++ libflux/go/libflux/buildinfo.gen.go
@@ -61,8 +61,8 @@ var sourceHashes = map[string]string{
 	"libflux/flux/Cargo.toml":                                                                     "f9f73d9176225e75b2bf658dce4b1e4294db8a451b96ba28323afad4840cd54b",
 	"libflux/flux/FLUXDOC.md":                                                                     "92e6dd8043bd87b4924e09aa28fb5346630aee1214de28ea2c8fc0687cad0785",
 	"libflux/flux/build.rs":                                                                       "eea9cfad417042bdd0908c5f7bcf2f004ee70031d973292e2cf54fab526a1807",
-	"libflux/flux/src/cffi.rs":                                                                    "b3d57670725039f848c414e0f796c435599c690bef7e6c8f7f4d5a3cc1373401",
-	"libflux/flux/src/lib.rs":                                                                     "32d305d27a04cd061b796c0f70bedbae9cb4fcdbcc7d08505c74cf9c22adc953",
+	"libflux/flux/src/cffi.rs":                                                                    "3141641d1f4a946d3bca4cd4bf5044fcaeed037293bf16fb45219130284477d2",
+	"libflux/flux/src/lib.rs":                                                                     "a1a0d74bf8c7f7d250c37b4442d7e1552f20cb11dc4921217958a885e42d3d5f",
 	"libflux/flux/templates/base.html":                                                            "a818747b9621828bb96b94291c60922db54052bbe35d5e354f8e589d2a4ebd02",
 	"libflux/flux/templates/home.html":                                                            "f9927514dd42ca7271b4817ad1ca33ec79c03a77a783581b4dcafabd246ebf3f",
 	"libflux/flux/templates/package.html":                                                         "635e1d638ad1a430f3894ff0a458572ca3c4dc6b9cec2c57fe771443849e1402",
